fix(overlay): make the on-demand-reconcile regression guard real, split connection.go under the length cap
TestOnDemandReconcileRacingDisconnectAnswersModeNotOverlay used a real,
completed Disconnect, which flips workspace.x_sor_mode to native in the
same transaction as the purge — so requireOverlayMode always returned
first and neither the ErrConnectionGone mapping nor the WithFence() call
was ever exercised, despite the test's own doc comment claiming both.
The test now restores only workspace.x_sor_mode/x_incumbent via raw SQL
after a genuine Disconnect (simulating a TTL-caching mode dispatcher
serving a stale "overlay" read), forcing the call through the mode gate
into the fenced write. Verified by falsification: deleting the
ErrConnectionGone mapping fails it, and dropping WithFence() fails it too
(confirmed both ways, then reverted).

Also splits connection.go (617 lines, over the 500-line cap Task 1's
reconnect flow pushed it past) the way the package already splits
Disconnect into teardown.go: reconnectConnection, deleteSupersededRef,
and existingConnectionStatus move verbatim into a new reconnect.go
sibling; cleanupOrphanedRef stays in connection.go since both the
fresh-insert and reconnect branches call it.
