+++ backend/internal/compose/jobs_overlay_worker_integration_test.go
@@ -402,16 +402,24 @@ func TestWorkerCleanStopsOnMidSweepDisconnect(t *testing.T) {
 	}
 }
 
-// TestOnDemandReconcileRacingDisconnectAnswersModeNotOverlay proves the
-// on-demand /overlay/reconcile boundary (overlay.Service.RequestSweep)
-// translates a disconnected workspace's sentinel into the same
-// ErrModeNotOverlay a workspace with no active connection already gets —
-// never an opaque 500 — and that the request leaves no overlay_sync_state
-// row behind: a sweep request racing a teardown must not repopulate what
-// the purge already removed. This is the regression guard for the P1 the
-// review found: RequestSweep must run against the FENCED store
-// (MirrorStore.WithFence), or a request racing a disconnect would silently
-// re-insert the sync-state row the teardown purged.
+// TestOnDemandReconcileRacingDisconnectAnswersModeNotOverlay reproduces the
+// real race a TTL-caching mode dispatcher opens: another process's
+// Disconnect can already have committed (connection revoked,
+// overlay_sync_state purged, workspace flipped to native) while THIS
+// process is still serving a stale cached "overlay" read. After a genuine
+// Connect + Disconnect, it restores ONLY workspace.x_sor_mode/x_incumbent
+// via raw SQL — never incumbent_connection or overlay_sync_state, which
+// stay exactly as the teardown left them — so requireOverlayMode passes
+// and RequestSweep is forced through to the fenced write instead of being
+// turned away earlier by the mode gate. This is the actual regression
+// guard for the two P1s the review found: (1) RequestSweep must run
+// against the FENCED store (MirrorStore.WithFence), or this stale-mode
+// window would let it silently re-insert the overlay_sync_state row the
+// teardown purged; (2) the fence's ErrConnectionGone must be mapped to
+// apperrors.ErrModeNotOverlay before it can cross the wire, or this
+// answers an opaque 500 instead. Deleting either one independently fails
+// this test — verified by falsification while writing it, not merely
+// asserted here.
 func TestOnDemandReconcileRacingDisconnectAnswersModeNotOverlay(t *testing.T) {
 	e := integration.Setup(t)
 	vault := keyvault.NewMemory()
@@ -426,8 +434,21 @@ func TestOnDemandReconcileRacingDisconnectAnswersModeNotOverlay(t *testing.T) {
 		t.Fatalf("Disconnect: %v", err)
 	}
 
+	// Simulate the stale cached "overlay" mode read: restore ONLY
+	// workspace.x_sor_mode/x_incumbent, leaving incumbent_connection
+	// revoked and overlay_sync_state purged exactly as Disconnect left
+	// them — so the mode gate passes and the call reaches the fence.
+	if err := database.WithWorkspaceTx(adminCtx, e.Pool, func(tx pgx.Tx) error {
+		_, execErr := tx.Exec(adminCtx,
+			`UPDATE workspace SET x_sor_mode = 'overlay', x_incumbent = 'hubspot'
+			 WHERE id = NULLIF(current_setting('app.workspace_id', true), '')::uuid`)
+		return execErr
+	}); err != nil {
+		t.Fatalf("restoring the stale cached overlay mode: %v", err)
+	}
+
 	if err := svc.RequestSweep(adminCtx); !errors.Is(err, apperrors.ErrModeNotOverlay) {
-		t.Fatalf("RequestSweep on a disconnected workspace = %v, want apperrors.ErrModeNotOverlay (not an opaque 500)", err)
+		t.Fatalf("RequestSweep racing a disconnect = %v, want apperrors.ErrModeNotOverlay (not an opaque 500)", err)
 	}
 
 	var syncStateRows int
@@ -437,7 +458,7 @@ func TestOnDemandReconcileRacingDisconnectAnswersModeNotOverlay(t *testing.T) {
 		t.Fatalf("counting overlay_sync_state rows: %v", err)
 	}
 	if syncStateRows != 0 {
-		t.Errorf("overlay_sync_state has %d row(s) after a sweep request on a disconnected workspace, want 0 — the fence must not repopulate what the teardown purged", syncStateRows)
+		t.Errorf("overlay_sync_state has %d row(s) after a sweep request racing a disconnect, want 0 — the fence must not repopulate what the teardown purged", syncStateRows)
 	}
 }
